Implement Le parameter handling in select command
ISO 7816-4 section 7.1.1 specifies:

If the Le field contains only bytes set to '00', then all the bytes corresponding to the selection option should be
returned within the limit of 256 for a short Le field, or 65 536 for an extended Le field. If the Le field is absent,
i.e., for returning no file control information, then the response data field shall also be absent.

The original code ignores the Le value and sends the PIV APT regardless of what Le was specified.  This change
enables the following behavior:

1)  If no Le is sent, then status 0x9000 is sent to indicate the applet was properly selected.
2)  If Le is non-zero, it is converted to the corresponding Ne value.  If the APT length is
      less than Ne, it is returned.  If the length is larger than Ne, then 0x61xx is returned,
      indicating that more bytes may be read.