Implement Le parameter handling in select command
ISO 7816-4 section 7.1.1 specifies:

If the Le field contains only bytes set to '00', then all the bytes corresponding to the selection option should be
returned within the limit of 256 for a short Le field, or 65 536 for an extended Le field. If the Le field is absent,
i.e., for returning no file control information, then the response data field shall also be absent.

The original code ignores the Le value and sends the PIV APT regardless of what Le was specified.  This change 
enables the following behavior:

1)  If no Le is sent, then status 0x9000 is sent to indicate the applet was properly selected.
2)  If Le = 0x00 is sent, then the default PIV APT is returned.
3)  Any other Le value (0x01 - 0xFF) is treated as an error.